requirements.txt
libvirt-python
uuid
